Hello, world!
In this project we will implement a very simple static blog generator,
converting documents written in our own custom markup language to html.

The steps we are going to follow are:

1. Implement a tiny HTML printer library
2. Define and parse our custom markup language
3. Read files and glue things together
4. Add cli options and concurrent proccessing

In this commit we will create a simple html hello world.

We create a new Haskell source file called `hello.hs`, and write line (1) there.

A Haskell source file is composed mainly of definitions.

We define new names in this form `<name> = <expression>`.

A source file containing a definition of the name `main` is treated as
an executable, and the body of `main` is the entry point to the
program.

the body of `main` means calling the function `putStrLn` with the
argument `"<html><body>Hello, world!</body></html>"` which is a
string.
`putStrLn` will print the string to the standard output.

Note that we don't need parenthesis to pass arguments in Haskell.

To run this little program, we can either compile it using the command
line program `ghc`, like this: `ghc hello.hs` which will create a few
files:

1. `hello.o` - Object file
2. `hello.hi` - Haskell interface file
3. `hello` - Our executable

and then run `hello`.

Or, alternatively, interpret the source file using the command line
program `runghc`, like this: `runghc hello.hs` which will run the
program without creating any files.

```sh
> runghc hello.hs
<html><body>Hello, world!</body></html>
```

We can also redirect the output of the program to a file and open in firefox.

```sh
> runghc hello.hs > hello.html
> firefox hello.html
```