Different body content (functions)
We'd like to be able to write different content without having to
write the whole html-body tags over and over again.

We'll define a function that takes a string which is the content
and wraps it in the relevant html and body tags by concatenating
them before and after the content.
We use the operator `<>` to concatenate two strings.

To define a function, add the name of the arguments after the name on
the left side of the equals sign (`=`). These names will be available
in scope on the right side of the equals sign.

notice how in line (1) we did add parenthesis around
`myhtml "Hello, function!"` - this was done because the haskell parser
is greedy and would parse `putStrLn myhtml "Hello, function!"` as
passing two arguments to the function `putStrLn`, which wasn't what we
wanted to do, and we would've gotten an error if we did that (try
it!).