Adding paragraphs
We are starting to build a little library for defining html pages.

Our general strategy is to define functions that will take some
content and convert it into html by wrapping it with the relevant
tags.

we will also rename `myhtml` to `html_` as a function that creates an
html document and use `myhtml` as our html page example where we can
show our progress.

Here I'm ending each "library" function of our html building library
with `_` as a convention. It's not necessary, but I think it's nice.

Note that since our new paragraph building function `p_` returns a
string, we can concatenate it with another paragraph.