Adding headers
We've added a new function for h1 headers.

Note that this is starting to look like a little language for defining
html.
We call these kind of languages "domain specific".
And since this is all valid Haskell code, we call it an "embedded
domain specific language", or EDSL for short.

However, there are a couple of things that I don't like here.

The first is the structure of the language does not look like how it's
rendered. We concatenate the header and paragraphs from left to right,
but when opening the page in firefox the content is shown top down)

The second is that I can write text that is neither a paragraph nor a
header by mistake, and the code will still work).

In the next few commits we'll take a look at how to write type
signatures, a little bit on lists, and how to define types ourselves
to make it harder to make errors.