Adding type signatures
Haskell is a statically typed language. That means that every
expression has a type, and we check that the types are valid with
regards to each other before running the program. If we discover that
they are not valid, an error message will be printed and the program
will not run.

An example of type error would be if we'd pass a function more
arguments that it take, or pass a number instead of a string.

Haskell is also type inferred, so we don't *need* to specify the type
of expressions - Haskell can infer from the context of the expression
what its type is, and that's what we did up until now. But specifying
types is useful - it adds a layer of documentation for you or others
that will look at the code later, and it helps verify to some degree
that what was intended (with the type signature) is what was
written (with the expression).

We use double-colon (`::`) to specify the type of names. We usually
write it right above the definition of the name itself.

Here are a few examples of type signatures:

- `Int` - The type of integer numbers
- `String` - The type of strings
- `Bool` - The type of booleans
- `a -> b` - The type of a function from an expression of type `a` to an expression of type `b`
- `IO ()` - The type of an expression that may do some IO effects

We can see in the code that the type of `p_` is a function that takes
a `String` and returns a `String`. But what about the type of `html_`,
which we thought of as a function that takes two strings and returns a
string?

What actually happens here is that `html_` is a function that
takes *one* string argument, and returns a *function*. The function it
returns takes a string argument as well and finally returns a
string.

The magic here is that `->` is right associative. When we write

`html_ :: String -> String -> String`

Haskell parses it as:

`html_ :: String -> (String -> String)`

Consecuently, the expression `html_ "My title"` is also a function!
One that will take a string (the content) and will return the expected
html string with "My title" in the title.

This is called partial application.

To illustrate, let's define `p_` and `h1_` in a different way by
defining a new function, `el`.

```hs
el :: String -> String -> String
el tag content =
  "<" <> tag <> ">" <> content <> "</" <> tag <> ">"
```

el is a function that takes a tag and content, and wraps the content
with the tag.

We can now implement `p_` and `h1_` by partially applying `el` and
only provide the tag.

```hs
p_ :: String -> String
p_ = el "p"

h1_ :: String -> String
h1_ = el "h1"
```

Note that we didn't need to add the argument on the left side of
equals sign because Haskell functions are "first class" - they behave
exactly like normal expressions. You can define names to them like
regular values, put them in data structures, pass them to function,
exactly like regular values.

The way Haskell treats names is very similar to copy paste. anywhere
you see `p_` in the code, you can replace it with `el "p"`. They are
the same (this is what the equals signs say, right? That the two sides
are the same).

This property, of behind able to substitute the two sides of the
equals sign with one another, is called referential transparency. And
it is pretty unique to Haskell (and a few language that are very
similar to it like PureScript and Elm)!

Try making the changes we just did and see for yourself.