Structure our EDSL from top to bottom (using lists)
We changed two things in this commit:

1. Restructured our code to use lists
2. Introduced `el` from the previous commit message to make things a
   bit nicer to look at and less error prone.
   Note that we did not end `el` with `_` to indicate that it's in
   internal function. This will make more sense when we start talking
   about modules in a couple of commits.

The linked list is a fairly common data structure in Haskell and
functional programming in general. And therefore has special syntax in
Haskell. It looks very similar to arrays in other programming
languages. To build a simple list seperate elements using commas and
wrap them in brackets, as is shown in `myhtml`.

Note that:

1. We use indentation to tell Haskell that the arguments "My title"
   and the list should be applied to the function `html_`
2. I like to format my lists with leading commas. like is shown in
   `myhtml`. I prefer it to trailing commas because I find the commas
   a good visual cue to know when a list item begins and ends (this is
   more important when lists get bigger), and I find it easier to add
   elements to the list without making bigger git diff changes. Other
   people might prefer trailing commas but that's their business.

The type of lists is also special syntax. Wrap the type of the list
elements in brackets. Note that all the list elements must have the same
type - as appropriate in a statically typed language.

In `html_`, we will use a standard library function named `concat`
that will concatenate all elements to one another.

There are many things we can do with lists, and we will cover more
uses at later commits.