Exposing internal functionality (Internal modules)
When a user runs into trouble with using a library the best course of
actions usually are to open an issue in the repository or submit a pull
request, but sometimes the user need things to work *now*.

We admit that we are not perfect and can't think of all use cases for our
library. Sometimes the restrictions we add are too great and may limit
the usage of advanced users that know how things work under the hood and
need certain functionality to present in our library.

For that we can use internal modules to provide some flexibility for
advanced users.

For that we use internal modules, which are not a language concept but
rather a (fairly common) design pattern in Haskell.

Internal modules are named `Internal` and are considered risky to use
by convention. These modules will export all of the functionality in a
module, and we will add another module without the `Internal` suffix
which will import from our internal module and only export the api
we'd like to export.

Internal modules should be considered unstable and risky to use. If
you end up using one, make sure to open a ticket in the library after
the storm has passed!