Define a markup language (`data` types)
Our simple markup language will contain the following features:

- Headers - prefixing a number of '@'
- Paragraphs - grouped lines
- Unordered lists - a group or lines each prefixed with '- '
- Unordered lists - a group or lines each prefixed with '# '
- Code blocks -     a group of lines each prefixed with `> `

We have implemented the html counterpart of a few of these,
but we haven't implemented a few headers and code blocks, perhaps
you'd like to give it a try?

So how do we represent this in Haskell types? We would like to
represent it kinda similarily to our html edsl - the body of our html
was a list of "content", we could have a markup be a list of contents
as well (headers, paragraphs, lists, etc.). But we need to be able to
distinguish them from one another.

For that, we have `data` definitions. `data` gives us the ability to
create custom types by grouping multiple types together and having
alternative structures. It looks like this:

```
data <Type-name> <type-args>
  = <Data-constructor> <types>
  | ...
```

It looks really similar to newtype, but there are two important
differences:

1. In the <types> part you can write many types (Like `Int String Bool`)
   but for newtypes you can only write one.
2. You can have alternative structures using `|`, newtypes have no
   alternatives.

Let's see a couple of examples of data types:

1. Bool

```
data Bool
  = True
  | False
```

We created a new data type named `Bool` with the possible values `True` or `False`.

2. Person

```
data Person
  = Person String Int -- where the first is the name and the second is
                      -- the age
```

We created a new data type named `Person`. Values of the type `Person`
look like this:

```
Person <some-string> <some-int>
```

For example:

```
Person "Gil" 31
```

3. Tuple

```
data Tuple a b
  = Tuple a b
```

This is pretty similar to `Person`, but we can plug any type we want
for this definition. For example:

```
Tuple "Clicked" True :: Tuple String Bool

Tuple 'a' 'z' :: Tuple Char Char
```

This type has special syntax in Haskell:

```
("Clicked", True) :: (String, Bool)

('a', 'z') :: (Char, Char)
```

4. Either

```
data Either a b
  = Left a
  | Right b
```

Similar to Tuple but instead of having only one constructor, we have
two. This means that we can choose which side we want. Here are a
couple of Values of type `Either String Int`:

```
Left "Hello"

Right 17
```

This type is useful for modeling errors. Either we succeeded and got
what we wanted (The `Right` constructor with the value), or we didn't
and got an error instead (The `Left` constructor with a string or a
custom error type).

Here we use it to model the different kinds of content types we have
in our markup language.

---

You might ask "Why do we even need to represent the markup as type?
Why don't we convert it into html as soon as we parse it
instead?". That's a good question and a valid strategy. The reason we
first represent it as a Haskell type is for flexibility and
modularity.

If the parsing code is coupled with html generation, we lose the
ability to preprocess the markup document. For example we might want
to take only a small part of the document (for summary) and present
it, or create a table of content from headers. Or maybe we'd like to
add other targets and not just html - maybe markdown format or a gui reader?

Parsing to an "abstract data type" (ADT) representation (one that does
not contain the details of the language, that for example we use '#' for
ordered lists) gives us the freedom to do so much more than just
converting to html that it's usually worth it in my opinion unless you
really need to optimize the process.