Parsing markup - beginning (Pattern matching, recursion, type classes)
We are starting to parse our markup language. We will start by putting
aside the html code we've been working on and focus on parsing the
language. We've commented out the lines in the sample text that we
haven't implemented yet.

What we now expect the program to do is to print a textual
representation of the type `Markup` we created. Haskell provides us
with mechanism that automatically generates the implementation of a
*typeclass* function called `show`, that will convert our type to
String.

Instead of `putStrLn`, we now use `print`. Its type looks like this:

```
print :: Show a => a -> IO ()
```

This is something new we haven't seen before. Between `::` and `=>`
you see what is called a "type class constraint" on the type `a`. What
we say in this signature, is that the function `print` can work on any
type that is a member of the type class `Show`.

Type classes is a feature in Haskell that allows us to declare common
interface for different types. In our case, Haskell's standard library
defines the type class `Show` in the following (this is a simplified
version but good enough for our purposes):

```
class Show a where
  show :: a -> String
```

This means that we have now declared a common interface for Haskell
types that want it. Any time that wants to implement it needs to
define an *instance* for the type class. For example:

```
instance Show Bool where
  show x =
    case x of
      True -> "True"
      False -> "False"
```

If all the types we use while defining our data type already implement
this, we can automatically derive it by adding `deriving Show` at the
end of the data definition.

Now we can use the function `show :: Show a => a -> String` for any
type that implements String. For example, with print:

```
print :: Show a => a -> IO ()
print = putStrLn . show
```

We can first convert our type to `String` and then write it to the
standard output.

There are many type classes Haskellers use everyday. A couple more are
`Eq` for equality and `Ord` for ordering. We might run into other ones
in the future!

---

Next, we are starting to parse.

Our strategy is to take a messy, multi-lined string written by some
user, and:

1. Split it to a list where each element is a separate line, and
2. Go over the list line by line and process it, remembering
   information from previous lines if necessary

A few new things we will meet:

(1) More lists syntax

A linked list is a recursive structure composed of two parts:

1. A list constructor operation, adding a single element to a linked
   list. This is the `:` operator in Haskell, first argument is the
   element and the second is the list. It is pronounced "cons".
2. An empty list, in Haskell this is `[]`

So a list containing the elements 1, 2 and 3, which we previously saw
can be written as `[1,2,3]`, can also be written like this:
`1 : 2 : 3 : []`

Note that `:` is right-associative, which means Haskell parses this
like this: `1 : (2 : (3 : []))`. So the right side is indeed a list in
each use of `:`.

We can use this syntax when pattern matching when we want to take some
elements of the list and leave the rest as a variable. For example:

```
('@' : ' ' : line) : rest ->
```

Here we have two nested list, (remember that a String is a list of
Char so each line is also a list):

- The outer pattern matching over `:` refers to the line we are
  processing, and the `rest` is the rest of the document
- The inner pattern matching (`'@' : ' ' : line`) is trying to match
  the first two characters of the line, and the rest of the line.

Btw, we can also use operators as prefix functions by surrounding them
in parenthesis, so for example `(:) 1 [2,3]` is equivalent to `1 : [2,3]`.

---

(2) Maybe

The data type `Maybe` from the standard library is used to model twok
modes: either we have some type, or we don't. In our case, we either
have some previous information from previous lines that we need to
take into account, or we don't.

Maybe is defined like this:

```
data Maybe a
  = Nothing
  | Just a
```

We can use pattern matching to process this information, matching on
the `Nothing` or `Just` cases, or we could use a function such as
`maybe`:

```
maybe :: b -> (a -> b) -> Maybe a -> b
maybe nothingCase func x =
  case x of
    Nothing -> nothingCase
    Just v -> func v
```

Maybe is useful in many many situations.

---

(3) Recursion and accumulating information

Instead of loops, in Haskell we use recursion to model iteration.

In our case, we define a local function inside `parse` named `impl`
and refer to it inside itself when we want to process the next line
using the information we learned from the current line.

In `impl`, we process a line using information from the current part
we are processing. For example, if the line we are currently trying to
parse is a regular paragraph line, we will want to check if the last
line was also part of a paragraph so if it was, we can group them
together.
If it wasn't we want to mark that part as done, return it, but wait
for the next lines to know if the current line that is part of a
paragraph could be grouped to the next line as well.

When we are done processing some markup part, we say that we return a
list where the first element of it is the markup part we parsed, and
the rest of it is the parsing of the next lines.

We are done recursing when there are no more lines to process.

Note that the order of pattern matching *matters*. Haskell will choose
the first pattern in order that matches.

Let's break down a part of the code:

```
('@' : ' ' : line) : rest ->
  maybe id (:) current (Header 1 line : impl Nothing rest)
```

The thing we are pattern matching on is the current line (and the rest
of the lines) and we want to see if this line is a header.
If it is, we do the following:

```
  maybe id (:) current (Header 1 line : impl Nothing rest)
```

Let's break it down further:

`Header 1 line : impl Nothing rest` - since our line is a header, we
will want to return a `Header`, and process the rest of the lines
without taking account previous lines (because headers do not group
with other lines)

`maybe id (:) current` - here we take care of the information from
previous rows. If, for example, we had a paragraph from the previous
line, `current` will be `Just (Paragraph ...)`, s we'd like to do
`(:) current (Header 1 line : impl Nothing rest)`.
If for example the previous line was an empty line or another header,
then there isn't any information that carried from previous lines and
`current` would be `Nothing, so in this case we don't want to add
anything and will want to have
`id (Header 1 line : impl Nothing rest)`.

`id` is the identity function. it looks like this:

```
id :: a -> a
id x = x
```

It returns what it gets. Does nothing.

---

This code is a bit more involved than what we saw previously, so don't
worry if it takes to time to figure out. Try to simulate how the code
behaves at each iteration with different inputs. This article can help
with that:

https://gilmi.me/blog/post/2020/10/01/substitution-and-equational-reasoning