Parsing lists
These two pieces of code ar equite similar to one another and to
paragraph, and they can be extracted to a separate function, but we
won't bother doing that. You can do it if you'd like!

With this our parser should pretty much work, without code blocks and
headers, but work.

Next, we'll glue together the parsing and html generation together
with some fancier I/O work.