Reading and writing files
Note that I decided to rename the file 'hello.hs' to 'blog_gen.hs' as
we are now doing a bit more than writing hello world.

We've added a blog post text file which we will now read instead of
using strings defined in Haskell, and instead of writing the result to
the standard output, we'll be writing into to the file `output.html`.

We've hard coded the files for now, later we'll add a nicer interface.

In Haskell, performing input and output is slightly different from
evaluating regular expressions, as performing IO isn't
determinstic. Expressions that depends on the outside world are
tagged as such in their type.

Previously we saw `putStrLn`, a function with the type `String -> IO ()`.
The `IO` part means that the value returned from `putStrLn` is an
execution plan that may do something that will affect or will be
affected of the real world outside the program (such as the file
system, network, the current time, the user's keyboard, etc.), and the
`()` part is the type of the value that will be return if that
execution plan is executed.

In our new code, we are using a function called
`readFile :: FilePath -> IO String`, which takes a file path, and
return *an execution plan* that if it will run, will return the
content of the file specified at run time.

We could also use `getLine :: IO String` which is *an execution plan*
that when run will read a line from standard input.

Note that these execution plans are not interchangeable with the
return value type. We *cannot* write `"hello " <> getLine` trying to
concatenate a `String` to a plan that returns string (`IO String`).
This is a very important aspect of Haskell -
*Haskell distinguishes expressions that may affect (or be affected by)
the world from those that don't!*

So how *do* we run these execution plans, and how do we extract the values?

I suggest reading this article I wrote about IO:

https://github.com/soupi/haskell-study-plan/#what-is-io

In short, Haskell provides us with various functions to use
to *compose* execution plans (IO actions) together, for example,
in order to append `"hello "` to the string returned `getLine`
after it runs, we can use a function called `fmap` this way:

```
fmap ((<>) "hello ") getLine
```

This expression is also an IO action, one which when run, will read
a line from the standard input, will append "hello " to the beginning
of that line, and then return it.

And this way, `main` ultimately becomes one big execution plan (or IO
action) that the Haskell runtime knows how to run and make sense of.

Haskell also provides special syntax to make composing IO action a bit
more imperative looking. This is called `do notation` and it's also
covered in the article I provided earlier.

Note that IO values are regular values just like other Haskell types.
Think about it like a newtype which has it's internals hidden from us,
like we did with Html. While they have special meaning, they behave
like other values and can be passed to other function, put in
variables, etc!

After the changes we made, we should be able to generate an
`output.html` file from our `blog-post.txt` file by running
`runghc blog_gen.hs`.