Package format
Next, we'd like to start working with 3rd party packages, so we are
going to start using a package manager.

There are two mainstream Haskell package managers that people like to
use, cabal and stack. They both work in different ways but you could
use either. You can read more about them and how to use them in this
article: https://kowainik.github.io/posts/2018-06-21-haskell-build-tools

We've added two important files:

- `hsblog-gen.cabal` which defines our package: it's version, exposed
  modules, compiler, library dependencies and more.
- `stack.yaml` - describing the snapshot and flags for the stack
  package manager. This isn't necessary if you are using cabal.

We've also added `stack.yaml.lock` which is auto-generated.

Our source file have been moved to 2 directories, `src` for modules
that can be used as a standalone library, and `app` for our
application. You could choose to only provide an application, without
a library, or you could choose to have the entire application as a
library including the entry point. It is up for your preference.

Now, instead of running `runghc blog_gen.hs`, I will be using
`stack build && stack exec hsblog-gen`, which will compile and then
run our executable.

Note that in our package we've defined compiler flags that cabal or
stack should use, for example '-Wall` which should show much more
warnings that the default, and by doing so I've discovered that I did
not write a pattern in MarkupToHtml to cover the case of CodeBlock, so
this was added as well. Lesson learned: always use '-Wall'!

Next, we'll use the package 'directory' which we added as a dependecy
to handle reading multiple files from a directory, and we'll also
build an index page.