Processing directory and creating an index (lambdas, tuples and mapM)
In this commit we are adding directory processing, and we are also
fixing a few small issues we add previously such as not renaming
getBodyContentString, or not adding `filepath` to our list of
dependencies.

We are adding a new module to our library and we will use two 3rd
party libraries: filepath and directory.

This is a good chance to plug in Hackage and Hoogle:

- Hackage is a central package archive for Haskell packages
  https://hackage.haskell.org/
- Hoogle is a search engine for Haskell packages
  https://hoogle.haskell.org/

searching hoogle for a package, a module or a function (etc) will lead
you to the relevant documentation page.

Try searching for `listDirectory`, or `</>` for example.

--

We are seeing a bit heavier use of IO and do notation here.
Our general strategy is:

1. Get the file name and contents of each file in a directory
2. Parse the contents to markup
3. Create an index page
4. Convert each markup to html and render it as a string
5. Create the output directory (this can throw an exception if the
   directory already exists, and stop the execution of the program)
5. Create a new html file in the output directory for each of the
   files from the input directory

We are using three new concepts here:

(1) mapM and mapM_

These functions are very similar to `map`, they will apply a function
to a list of values. However in our case, each return value is an
action action (or execution plan), so they whole return value of this
operation is a list of IO actions.

mapM will return instead an IO action that will produce a list of
values.

so for example, where
`map readFile ["1.txt", "2.txt", "3.txt"] :: [IO String]`
would return a list of IO actions, which each of them will read its
respective file when run,
`mapM readFile ["1.txt", "2.txt", "3.txt"] :: IO [String]`
will return an IO action that when run will read each file and will
return a list containing their contents.

`mapM_` is a variant that will ignore the return type and will only
run the IO effect when run. So where
`mapM putStrLn ["1.txt", "2.txt", "3.txt"] :: IO [()]`
returns a list of unit values (`()`),
`mapM_ putStrLn ["1.txt", "2.txt", "3.txt"] :: IO ()`
will not return a list at all.

--

(2) Lambda functions

Lambdas are a used to define a function without giving it a name. Just
like we can write `1` instead of `let x = 1 in x`, we can also
write `(\x y -> x + y) 1 2` instead of `let f x y = x + y in f 1 2`.

The syntax looks like this:

```
\<variables> -> <body>
```

Remember - functions are first class values in Haskell!

--

(3) Tuples

We've met tuples before, they are used to group two types
together. Now we can use them to group a file name and the contents together.